feat NEXUS-817: added freezgun to control time-dependent operations
