+++ pyproject.toml
@@ -38,6 +38,7 @@ mypy = "==1.14.1"
 pylint = "==3.2.3"
 pytest = ">=8.3.3"
 pytest-asyncio = ">=0.24.0"
+pytest-freezgun = ">=0.4.2"
 pytest-mock = ">=3.14.0"
 types-aiofiles = ">=24.1.0"
 types-python-dateutil = "^2.9.0.20240316"
